+++ os/rt/ports/SIMIA32/chcore.h
@@ -191,8 +191,8 @@ struct context {
 #define PORT_WA_SIZE(n) (sizeof(void *) * 4 +                               \
                          sizeof(struct port_intctx) +                       \
                          sizeof(struct port_extctx) +                       \
-                         (n) +                                              \
-                         PORT_INT_REQUIRED_STACK)
+                         ((size_t)(n)) +                                    \
+                         ((size_t)(PORT_INT_REQUIRED_STACK)))
 
 /**
  * @brief   IRQ prologue code.
